backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.7.1
pytest==8.2.2
httpx==0.28.1
ruff==0.4.4
passlib[bcrypt]==1.7.4
bcrypt<4.1.0
pyjwt==2.8.0
requests==2.32.2